requirements.txt
aniso8601==7.0.0
asgiref==3.4.1; python_version >= "3.8"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.10; python_full_version >= "3.6.0" and python_version >= "3"
cryptography==36.0.1; python_version >= "3.6"
defusedxml==0.7.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
django-allauth==0.47.0
django-crispy-forms==1.13.0; python_version >= "3.6"
django-debug-toolbar==3.2.4; python_version >= "3.6"
django-extensions==3.1.5; python_version >= "3.6"
django-filter==21.1; python_version >= "3.6"
django-storages==1.12.3
django==4.0.1; python_version >= "3.8"
djangorestframework==3.13.1; python_version >= "3.6"
graphene-django==2.15.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
importlib-metadata==4.10.1; python_version < "3.10" and python_version >= "3.7"
markdown==3.3.6; python_version >= "3.6"
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
postgres==4.0
promise==2.3
psycopg2==2.9.3; sys_platform == "win32" and python_version >= "3.6"
psycopg2-binary==2.9.3
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pyjwt==2.3.0; python_version >= "3.6"
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3; python_version >= "3.6"
requests-oauthlib==1.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.27.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
rx==1.6.1
singledispatch==3.7.0; python_version >= "2.6"
six==1.16.0; 
sqlparse==0.4.2; python_version >= "3.8"
text-unidecode==1.3
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8"
urllib3==1.26.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
zipp==3.7.0; python_version < "3.10" and python_version >= "3.7"
